update openapi logic
