+++ pkg/microservice/aslan/core/environment/handler/openapi.go
@@ -1036,38 +1036,6 @@ func OpenAPIUpdateGlobalVariables(c *gin.Context) {
 	ctx.Err = service.OpenAPIUpdateGlobalVariables(args, ctx.UserName, ctx.RequestID, projectName, envName, ctx.Logger)
 }
 
-func OpenAPIUpdateRegistry(c *gin.Context) {
-	ctx := internalhandler.NewContext(c)
-	defer func() { internalhandler.JSONResponse(c, ctx) }()
-
-	args := new(UpdateProductRegistryRequest)
-	data, err := c.GetRawData()
-	if err != nil {
-		ctx.Err = e.ErrInvalidParam.AddDesc(err.Error())
-		return
-	}
-	if err = json.Unmarshal(data, args); err != nil {
-		ctx.Err = e.ErrInvalidParam.AddDesc(err.Error())
-		return
-	}
-
-	projectName, envName, err := generalOpenAPIRequestValidate(c)
-	if err != nil {
-		ctx.Err = e.ErrInvalidParam.AddErr(err)
-		return
-	}
-
-	internalhandler.InsertDetailedOperationLog(c, ctx.UserName, projectName, setting.OperationSceneEnv, "(OpenAPI)"+"更新", "环境-镜像仓库", envName, string(data), ctx.Logger, envName)
-
-	err = commonutil.CheckZadigXLicenseStatus()
-	if err != nil {
-		ctx.Err = err
-		return
-	}
-
-	ctx.Err = service.UpdateProductRegistry(envName, projectName, args.RegistryID, ctx.Logger)
-}
-
 func OpenAPIUpdateProductionYamlServices(c *gin.Context) {
 	ctx, err := internalhandler.NewContextWithAuthorization(c)
 	defer func() { internalhandler.JSONResponse(c, ctx) }()
