+++ pkg/microservice/aslan/core/system/handler/router.go
@@ -426,6 +426,7 @@ func (*OpenAPIRouter) Inject(router *gin.RouterGroup) {
 		reg.POST("", OpenAPICreateRegistry)
 		reg.GET("", OpenAPIListRegistry)
 		reg.GET(":id", OpenAPIGetRegistry)
+		reg.PUT(":id", OpenAPIUpdateRegistry)
 	}
 
 	cluster := router.Group("cluster")
